Launch and more

+++ packages/create-django-app/scripts/create-django-app.sh
@@ -7,28 +7,25 @@ SCRIPT_PATH="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 rm -rf "$SCRIPT_PATH/.venv"
 python3 -m venv "$SCRIPT_PATH/.venv"
 
-PROJECT_NAME="testproject"
+PROJECT_NAME=$1
+
+if [ -z ${PROJECT_NAME+x} ]; then
+    echo "You must pass in --name"
+    exit
+fi
 
 "$SCRIPT_PATH/.venv/bin/pip" install Django==4.0.1
 rm -rf $PROJECT_NAME
 "$SCRIPT_PATH/.venv/bin/django-admin" startproject server
 mv server $PROJECT_NAME
-# cp "$SCRIPT_PATH/shell.nix" "$PROJECT_NAME/shell.nix"
-# cp "$SCRIPT_PATH/requirements.txt" "$PROJECT_NAME/requirements.txt"
-sed -i 's/parent.parent/parent.parent.parent/g' "$PROJECT_NAME/server/settings.py"
 mkdir "$PROJECT_NAME/server/settings"
+rm "$PROJECT_NAME/server/settings.py"
 
 # TODO: remove me.
 mkdir -p "$PROJECT_NAME/static"
 
-mv "$PROJECT_NAME/server/settings.py" "$PROJECT_NAME/server/settings/common.py"
-# cp "$SCRIPT_PATH/localhost.py" "$PROJECT_NAME/server/settings/localhost.py"
 ln -s  localhost.py "$PROJECT_NAME/server/settings/__init__.py"
-# mkdir "$PROJECT_NAME/client"
-# cp "$SCRIPT_PATH/index.tsx.template" "$PROJECT_NAME/client/index.tsx"
-# cp "$SCRIPT_PATH/tsconfig.json.template" "$PROJECT_NAME/tsconfig.json"
-# cp "$SCRIPT_PATH/.babelrc.json.template" "$PROJECT_NAME/.babelrc.json"
-# cp "$SCRIPT_PATH/.babelrc.json.template" "$PROJECT_NAME/.babelrc.json"
+
 cp -RT "$SCRIPT_PATH/../template" "$PROJECT_NAME"
 cd "$PROJECT_NAME" || exit
 nix-shell --command "yarn init --yes && yarn add reactivated@0.20.1-a641"
